new_requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cached-property==1.5.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
config==0.5.1
cryptography==2.8
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-SSE==1.0.0
frozenlist==1.3.3
gevent==23.7.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.21.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
packaging==23.1
password-strength==0.0.3.post2
pkg_resources==0.0.0
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
python-dateutil==2.8.2
python-Levenshtein==0.21.1
pytz==2023.3
rapidfuzz==3.2.0
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
sqlalchemy-orm==1.2.10
sse==1.2
twilio==8.2.1
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
